+++ .env.example
@@ -1,1 +1,2 @@
 REACT_EDITOR=code
+NEXT_PUBLIC_API_URL=
