+++ package.json
@@ -27,6 +27,7 @@
     "slick-carousel": "^1.8.1",
     "tailwind-merge": "^2.5.2",
     "tailwindcss-animate": "^1.0.7",
+    "usehooks-ts": "^3.1.0",
     "yup": "^1.4.0",
     "zod": "^3.23.8",
     "zustand": "^4.5.5"
