+++ src/components/common/app-loading.tsx
@@ -8,10 +8,7 @@ import AppIcon from './app-icon'
 const AppLoading = memo(function AppLoading() {
   return (
     <div className="absolute left-0 top-0 z-50 flex h-full w-full items-center justify-center bg-white delay-1000 duration-1000">
-      <AppIcon
-        src="/svg/circle-loading.svg"
-        className="text-primary h-10 w-10 animate-spin md:h-20 md:w-20"
-      />
+      <AppIcon src="/svg/circle-loading.svg" className="h-10 w-10 animate-spin text-primary" />
     </div>
   )
 }, isEqual)
@@ -21,7 +18,7 @@ const LoadingIcon = memo(function LoadingIcon({ className }: { className?: strin
   return (
     <AppIcon
       src="/svg/circle-loading.svg"
-      className={cn('text-primary size-5 animate-spin', className)}
+      className={cn('size-5 animate-spin text-primary', className)}
     />
   )
 }, isEqual)
